Crate esp_ward

esp-ward

esp-ward is a Rust crate designed as a higher-level abstraction over esp-hal to simplify the usage of ESP32, ESP32S2, ESP32C3, ESP32C6, ESP32S3, and ESP32H2 chips with Rust. It provides common APIs, traits, and structs to interact with various peripherals such as GPIOs, I2C, and SPI devices.

This crate is targeted at developers new to the esp-rs ecosystem or those who prefer a simplified interface for common operations.

Features

Traits and structs for common peripheral interactions.

Easy configuration of SPI and I2C.

Predefined macros for common operations and setup routines.

Compatible with various ESP32 family chips.

Simplified Wi-Fi and MQTT features

Usage

To use esp-ward, include it as a dependency in your Cargo.toml and refer to the following examples to start interacting with your ESP device’s hardware features.

Quick Start

Here’s how you might initialize the system peripherals and configure I2C and SPI with default settings:

use esp_ward::{init_chip, take_periph, take_system};

let peripherals = take_periph!();
let system = take_system!(peripherals);
let (clocks, pins) = init_chip!(peripherals, system);
// Now you can use `clocks` and `pins` to interact with the peripherals

Example: Configuring I2C

use esp_ward::init_i2c_default;

let i2c = init_i2c_default!(peripherals, pins, clocks);
// Now `i2c` is ready to communicate with I2C devices

Example: Configuring SPI

use esp_ward::init_spi_default;

let spi = init_spi_default!(peripherals, pins, clocks);
// Now `spi` is ready to transfer data with SPI devices

Macros

This crate also provides several macros to ease the setup and usage of ESP peripherals.

take_periph

Takes the peripherals from the ESP board. This is typically one of the first steps in a Rust-ESP application.

take_system

Splits the SYSTEM peripheral into its constituent parts.

init_chip

Initializes the system clocks and IO pins.

init_i2c_default and init_i2c_custom

Initializes the I2C peripheral with either default or custom configurations.

init_spi_default and init_spi_custom

Initializes the SPI peripheral with either default or custom configurations.

init_wifi

Initializes Wi-Fi connection in async or non-async way - depending on your project

and more…

Contributing

Contributions to esp-ward are welcome. Check out the repository on GitHub to report issues or submit pull requests.

License

esp-ward is distributed under the terms of both the MIT license and the Apache License (Version 2.0).

See LICENSE-APACHE and LICENSE-MIT for details.

Modules

connectivity

Connectivity Features

display

Display Module

nb

Minimal and reusable non-blocking I/O layer

peripherals

Peripherals Module

Macros

create_joystick

Macro for creating a Joystick instance.

create_stack

Macro to create a network stack for WiFi communication.

get_ip

Macro to retrieve the IP configuration from the network stack.

get_timer

Macro to obtain a suitable timer based on the ESP device mod

init_chip

Initializes the system clocks and IO pins, providing the base setup required for any operation with peripherals.

init_i2c_custom

Initializes a custom I2C configuration, allowing for arbitrary SDA and SCL pins and frequency.

init_i2c_default

Initializes the default I2C configuration for the ESP board. Assumes the use of the standard I2C0 peripheral and “default” pin configuration. The rest of “default” functions

init_spi_custom

Initializes a custom SPI configuration, allowing for arbitrary CLK, MOSI, MISO, and CS pins and frequency.

init_spi_default

Initializes the default SPI configuration for the chip. Assumes the use of the standard SPI2 peripheral and default pin configuration.

init_wifi

Macro to initialize the WiFi interface with the given SSID and password in mqtt (or async) configuration. This macro configures the WiFi controller and initializes the WiFi interface. Example:

load_lp_code

Load code to be run on the LP/ULP core.

make_gpio_enum_dispatch_macro

Create an enum for erased GPIO pins, using the enum-dispatch pattern

prepare_alloc

Sets up a global allocator for heap memory, required for the alloc crate functionalities. This is essential for using heap-allocated data structures which are used, for example, for max7219 display. ATTENTION: MAKE SURE to use this prepare_alloc as a first function in your program if you’re using module which utilizes alloc. Modules like this will have a warning that you should use the alloc feature

prepare_buffers

Macro to prepare buffers with fixed sizes for MQTT communication.

take_periph

Takes the ESP peripherals. This should be one of the first steps in an ESP application, ensuring that the peripherals are properly acquired before use.

take_system

Splits the SYSTEM peripheral into its constituent parts. This macro is a convenience wrapper for quickly accessing system components.

wait

Pauses the execution for a specified number of milliseconds using a delay provider.

wait_wifi

Macro to wait until WiFi is connected in async variation Typically used after net_task async task call.

Structs

Clocks

Frozen clock frequencies

FullDuplexMode

Full-duplex operation

I2C

I2C peripheral container (I2C)

IO

General Purpose Input/Output driver

Peripherals

Pins

Spi

SPI peripheral driver

Enums

SpiMode

SPI modes

Traits

I2cInstance

I2C Peripheral Instance

InputPin

OutputPin

Peripheral

Trait for any type that can be used as a peripheral of type P.

SpiInstance

_embedded_dma_ReadBuffer

Trait for buffers that can be given to DMA for reading.

_embedded_dma_ReadTarget

Trait for Deref targets used by the blanket DmaReadBuffer impl.

_embedded_dma_Word

Trait for DMA word types used by the blanket DMA buffer impls.

_embedded_dma_WriteBuffer

Trait for buffers that can be given to DMA for writing.

_embedded_dma_WriteTarget

Trait for DerefMut targets used by the blanket DmaWriteBuffer impl.

_embedded_hal_Capture

Input capture

_embedded_hal_Pwm

Pulse Width Modulation

_embedded_hal_PwmPin

A single PWM channel / pin

_embedded_hal_Qei

Quadrature encoder interface

_embedded_hal_adc_OneShot

ADCs that sample on single channels per request, and do so at the time of the request.

_embedded_hal_blocking_delay_DelayMs

Millisecond delay

_embedded_hal_blocking_delay_DelayUs

Microsecond delay

_embedded_hal_blocking_i2c_Read

Blocking read

_embedded_hal_blocking_i2c_Write

Blocking write

_embedded_hal_blocking_i2c_WriteRead

Blocking write + read

_embedded_hal_blocking_rng_Read

Blocking read

_embedded_hal_blocking_serial_Write

Write half of a serial interface (blocking variant)

_embedded_hal_blocking_spi_Transfer

Blocking transfer

_embedded_hal_blocking_spi_Write

Blocking write

_embedded_hal_digital_InputPinDeprecated

Single digital input pin

_embedded_hal_digital_OutputPinDeprecated

Single digital push-pull output pin

_embedded_hal_digital_ToggleableOutputPinDeprecated

Output pin that can be toggled

_embedded_hal_digital_v2_InputPin

Single digital input pin

_embedded_hal_digital_v2_OutputPin

Single digital push-pull output pin

_embedded_hal_digital_v2_StatefulOutputPin

Push-pull output pin that can read its output state

_embedded_hal_digital_v2_ToggleableOutputPin

Output pin that can be toggled

_embedded_hal_serial_Read

Read half of a serial interface

_embedded_hal_serial_Write

Write half of a serial interface

_embedded_hal_spi_FullDuplex

Full duplex (master mode)

_embedded_hal_timer_CountDown

A count down timer

_embedded_hal_watchdog_Watchdog

Feeds an existing watchdog to ensure the processor isn’t reset. Sometimes commonly referred to as “kicking” or “refreshing”.

_embedded_hal_watchdog_WatchdogDisable

Disables a running watchdog timer so the processor won’t be reset.

_embedded_hal_watchdog_WatchdogEnable

Enables A watchdog timer to reset the processor if software is frozen or stalled.

_esp_hal_clock_Clock

_esp_hal_dma_DmaTransfer

Trait to be implemented for an in progress dma transfer.

_esp_hal_dma_DmaTransferRxTx

Trait to be implemented for an in progress dma transfer.

_esp_hal_gpio_InputPin

_esp_hal_gpio_OutputPin

_esp_hal_gpio_Pin

_esp_hal_i2c_Instance

I2C Peripheral Instance

_esp_hal_ledc_channel_ChannelHW

Channel HW interface

_esp_hal_ledc_channel_ChannelIFace

Channel interface

_esp_hal_ledc_timer_TimerHW

Interface for HW configuration of timer

_esp_hal_ledc_timer_TimerIFace

Interface for Timers

_esp_hal_system_SystemExt

Extension trait to split a SYSTEM/DPORT peripheral in independent logical parts

_esp_hal_timer_Instance

Timer peripheral instance

_esp_hal_timer_TimerGroupInstance

_esp_hal_uart_Instance

UART peripheral instance

_esp_hal_uart_UartPins

Pins used by the UART interface

_fugit_ExtU32

Extension trait for simple short-hands for u32 Durations

_fugit_ExtU64

Extension trait for simple short-hands for u64 Durations

_fugit_RateExtU32

Extension trait for simple short-hands for u32 Rate

_fugit_RateExtU64

Extension trait for simple short-hands for u64 Rate

Attribute Macros

entry

Attribute to declare the entry point of the program

interrupt

Marks a function as an interrupt handler

main

Creates a new executor instance and declares an application entry point spawning the corresponding function body as an async task.

ram

This attribute allows placing statics and functions into ram.