Module esp_ward::connectivity

Connectivity Features

This module provides functionality for initializing and managing WiFi connections on ESP devices, including MQTT messaging The creation of this functionality is inspired by the documentation and examples provided in esp-wifi (https://github.com/esp-rs/esp-wifi/tree/main)

Modules

mqtt

this module was written based on a basis of an analysis of an existing project from an esp-rs team member Juraj Sadel, that was shown at Espressif DevCon 2023. Available on: https://github.com/JurajSadel/esp32c3-no-std-async-mqtt-demo

wifi

The functions for dealing with time and timestamps were taken from experiments from my own project from the past - esp-clock-nostd (available on: https://github.com/playfulFence/esp-clock-nostd/tree/main)