Module esp_ward::connectivity::mqtt

this module was written based on a basis of an analysis of an existing project from an esp-rs team member Juraj Sadel, that was shown at Espressif DevCon 2023. Available on: https://github.com/JurajSadel/esp32c3-no-std-async-mqtt-demo

Macros

create_stack

Macro to create a network stack for WiFi communication.

get_ip

Macro to retrieve the IP configuration from the network stack.

prepare_buffers

Macro to prepare buffers with fixed sizes for MQTT communication.

wait_wifi

Macro to wait until WiFi is connected in async variation Typically used after net_task async task call.

Constants

HIVE_MQ_IP

Represents the IP address for the HiveMQ MQTT broker.

HIVE_MQ_PORT

Represents the port number for the HiveMQ MQTT broker.

Functions

mqtt_connect_custom

Establishes a custom MQTT connection with the specified parameters.

mqtt_connect_default

Establishes a default MQTT connection with predefined settings: HiveMQ broker

mqtt_receive

Waits for and receives a message from the subscribed MQTT topics. It handles reconnection in case of network errors.

mqtt_send

This function attempts to send the message to a specific MQTT topic and retries in case of network errors.

mqtt_subscribe

Subscribes to an MQTT topic.