Module esp_ward::display::ili9341

ILI9341 Display Driver

This module provides a driver for the ILI9341 LCD display using the SPI interface. It includes functionalities to interact with the display at a low level, such as setting pixels and writing strings, as well as high-level operations via the embedded-graphics library.

Structs

Ili9341Display

The Ili9341Display struct represents an ILI9341 display connected via SPI.

Constants

DEFAULT_STYLE_LARGE

DEFAULT_STYLE_MID

DEFAULT_STYLE_SMALL