Trait esp_ward::display::EGDisplay

pub trait EGDisplay {
    // Required methods
    fn write_to_segment(
        &mut self,
        segment: DisplaySegment,
        text: &str,
        font: MonoTextStyle<'_, Rgb565>
    );
    fn write_segment_name(
        &mut self,
        segment: DisplaySegment,
        name: &str,
        font: MonoTextStyle<'_, Rgb565>
    );
}

Extension of the Display trait to integrate with the embedded_graphics library.

Required Methods

fn write_to_segment( &mut self, segment: DisplaySegment, text: &str, font: MonoTextStyle<'_, Rgb565> )

Writes a string to a specific segment of the display using a specified font style.

fn write_segment_name( &mut self, segment: DisplaySegment, name: &str, font: MonoTextStyle<'_, Rgb565> )

Writes a section name to a specific segment of the display using a specified font style. This can be used to label parts of the display for better organization and readability.