Module esp_ward::peripherals::joystick

Joystick Module

Provides an interface to a 2-axis joystick with an integrated select button. This module is configured to use specific pins for the X and Y axes and assumes that the select button uses a digital input pin.

ATTENTION: THIS MODULE IS SUPPOSED TO BE USED ONLY WITH X-AXIS AND Y-AXIS CONNECTED TO DEFAULT PINS!!!

Macros

create_joystick

Macro for creating a Joystick instance.

Structs

Joystick

Represents a joystick with two axes and a select button.

Constants

ROUGH_THRESHOLD

A threshold value to interpret the joystick’s value in direction.