Struct esp_ward::peripherals::pir::PIRSensor

pub struct PIRSensor<PIN: InputPin> { /* private fields */ }

Represents a PIR motion sensor connected to a single digital input pin.

Implementations

impl<PIN: InputPin<Error = Infallible>> PIRSensor<PIN>

pub fn create_on_pins(pin: PIN) -> Self

Constructs a new PIRSensor with the given input pin.

Arguments

pin - The digital input pin connected to the PIR sensor.

Returns

A new PIRSensor instance.

pub fn motion_detected(&mut self) -> bool

Checks if motion has been detected by the sensor.

Returns

Returns true i