Struct esp_ward::I2C

pub struct I2C<'d, T> { /* private fields */ }

I2C peripheral container (I2C)

Implementations

impl<'d, T> I2C<'d, T>
where T: Instance,

pub fn new<SDA, SCL>( i2c: impl Peripheral<P = T> + 'd, sda: impl Peripheral<P = SDA> + 'd, scl: impl Peripheral<P = SCL> + 'd, frequency: Rate<u32, 1, 1>, clocks: &Clocks<'_> ) -> I2C<'d, T>
where SDA: OutputPin + InputPin, SCL: OutputPin + InputPin,

Create a new I2C instance This will enable the peripheral but the peripheral won’t get automatically disabled when this gets dropped.

pub fn new_with_timeout<SDA, SCL>( i2c: impl Peripheral<P = T> + 'd, sda: impl Peripheral<P = SDA> + 'd, scl: impl Peripheral<P = SCL> + 'd, frequency: Rate<u32, 1, 1>, clocks: &Clocks<'_>, timeout: Option<u32> ) -> I2C<'d, T>
where SDA: OutputPin + InputPin, SCL: OutputPin + InputPin,

Create a new I2C instance with a custom timeout value. This will enable the peripheral but the peripheral won’t get automatically disabled when this gets dropped.