Struct esp_ward::Peripherals

pub struct Peripherals {

    pub ADC1: ADC1,
    pub AES: AES,
    pub ASSIST_DEBUG: ASSIST_DEBUG,
    pub ATOMIC: ATOMIC,
    pub BT: BT,
    pub DMA: DMA,
    pub DS: DS,
    pub ECC: ECC,
    pub EFUSE: EFUSE,
    pub EXTMEM: EXTMEM,
    pub GPIO: GPIO,
    pub GPIO_SD: GPIO_SD,
    pub HINF: HINF,
    pub HMAC: HMAC,
    pub HP_APM: HP_APM,
    pub HP_SYS: HP_SYS,
    pub I2C0: I2C0,
    pub I2S0: I2S0,
    pub IEEE802154: IEEE802154,
    pub INTERRUPT_CORE0: INTERRUPT_CORE0,
    pub INTPRI: INTPRI,
    pub IO_MUX: IO_MUX,
    pub LEDC: LEDC,
    pub LPWR: LPWR,
    pub LP_CORE: LP_CORE,
    pub LP_PERI: LP_PERI,
    pub LP_ANA: LP_ANA,
    pub LP_AON: LP_AON,
    pub LP_APM: LP_APM,
    pub LP_APM0: LP_APM0,
    pub LP_I2C0: LP_I2C0,
    pub LP_I2C_ANA_MST: LP_I2C_ANA_MST,
    pub LP_IO: LP_IO,
    pub LP_TEE: LP_TEE,
    pub LP_TIMER: LP_TIMER,
    pub LP_UART: LP_UART,
    pub LP_WDT: LP_WDT,
    pub MCPWM0: MCPWM0,
    pub MEM_MONITOR: MEM_MONITOR,
    pub OTP_DEBUG: OTP_DEBUG,
    pub PARL_IO: PARL_IO,
    pub PAU: PAU,
    pub PCNT: PCNT,
    pub PMU: PMU,
    pub RMT: RMT,
    pub RNG: RNG,
    pub RSA: RSA,
    pub SHA: SHA,
    pub SLCHOST: SLCHOST,
    pub SOC_ETM: SOC_ETM,
    pub SPI0: SPI0,
    pub SPI1: SPI1,
    pub SPI2: SPI2,
    pub SYSTEM: SYSTEM,
    pub SYSTIMER: SYSTIMER,
    pub TEE: TEE,
    pub TIMG0: TIMG0,
    pub TIMG1: TIMG1,
    pub TRACE0: TRACE0,
    pub TWAI0: TWAI0,
    pub TWAI1: TWAI1,
    pub UART0: UART0,
    pub UART1: UART1,
    pub UHCI0: UHCI0,
    pub USB_DEVICE: USB_DEVICE,
    pub WIFI: WIFI,

}

Fields
ADC1: ADC1AES: AESASSIST_DEBUG: ASSIST_DEBUGATOMIC: ATOMICBT: BTDMA: DMADS: DSECC: ECCEFUSE: EFUSEEXTMEM: EXTMEMGPIO: GPIOGPIO_SD: GPIO_SDHINF: HINFHMAC: HMACHP_APM: HP_APMHP_SYS: HP_SYSI2C0: I2C0I2S0: I2S0IEEE802154: IEEE802154INTERRUPT_CORE0: INTERRUPT_CORE0INTPRI: INTPRIIO_MUX: IO_MUXLEDC: LEDCLPWR: LPWRLP_CORE: LP_CORELP_PERI: LP_PERILP_ANA: LP_ANALP_AON: LP_AONLP_APM: LP_APMLP_APM0: LP_APM0LP_I2C0: LP_I2C0LP_I2C_ANA_MST: LP_I2C_ANA_MSTLP_IO: LP_IOLP_TEE: LP_TEELP_TIMER: LP_TIMERLP_UART: LP_UARTLP_WDT: LP_WDTMCPWM0: MCPWM0MEM_MONITOR: MEM_MONITOROTP_DEBUG: OTP_DEBUGPARL_IO: PARL_IOPAU: PAUPCNT: PCNTPMU: PMURMT: RMTRNG: RNGRSA: RSASHA: SHASLCHOST: SLCHOSTSOC_ETM: SOC_ETMSPI0: SPI0SPI1: SPI1SPI2: SPI2SYSTEM: SYSTEMSYSTIMER: SYSTIMERTEE: TEETIMG0: TIMG0TIMG1: TIMG1TRACE0: TRACE0TWAI0: TWAI0TWAI1: TWAI1UART0: UART0UART1: UART1UHCI0: UHCI0USB_DEVICE: USB_DEVICEWIFI: WIFI
Implementations

impl Peripherals

pub fn take() -> Peripherals

Returns all the peripherals once

impl Peripherals

pub unsafe fn steal() -> Peripherals

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.