Trait esp_ward::_embedded_hal_watchdog_WatchdogEnable

pub trait _embedded_hal_watchdog_WatchdogEnable {
    type Time;

    // Required method
    fn start<T>(&mut self, period: T)
       where T: Into<Self::Time>;
}

Enables A watchdog timer to reset the processor if software is frozen or stalled.

Required Associated Types

type Time

Unit of time used by the watchdog

Required Methods

fn start<T>(&mut self, period: T)
where T: Into<Self::Time>,

Starts the watchdog with a given period, typically once this is done the watchdog needs to be kicked periodically or the processor is reset.

Object Safety

This trait is not object safe.